Fix cleanup() fails when base_dir does not exist

+++ src/testing/postgresql.py
@@ -205,7 +205,7 @@ class Postgresql(object):
 
     def cleanup(self):
         from shutil import rmtree
-        if self._use_tmpdir:
+        if self._use_tmpdir and os.path.exists(self.base_dir):
             rmtree(self.base_dir)
 
     def read_log(self):
